requirements.txt
streamlit==0.87.0
textblob==0.15.3
st-annotated-text==2.0.0
